+++ src/funcTest/groovy/info/solidsoft/gradle/pitest/functional/AcceptanceTestsInSeparateSubprojectFunctionalSpec.groovy
@@ -1,33 +1,16 @@
 package info.solidsoft.gradle.pitest.functional
 
 import nebula.test.functional.ExecutionResult
+import spock.lang.Ignore
 
 class AcceptanceTestsInSeparateSubprojectFunctionalSpec extends AbstractPitestFunctionalSpec {
 
+    @Ignore("Not compatible with Android Gradle plugin 3")
     def "should mutate production code in another subproject"() {
         given:
-            buildFile << """
-                apply plugin: 'com.android.library'
-                apply plugin: 'pl.droidsonroids.pitest'
-
-                android {
-                    buildToolsVersion '26.0.1'
-                    compileSdkVersion 26
-                    defaultConfig {
-                        minSdkVersion 10
-                        targetSdkVersion 26
-                    }
-                }
-                buildscript {
-                    repositories { mavenCentral() }
-                }
-                pitest {
-                    targetClasses = ['**']
-                }
-            """
             copyResources("testProjects/multiproject", "")
         when:
-            ExecutionResult result = runTasksSuccessfully('pitestRelease')
+            ExecutionResult result = runTasksSuccessfully(':itest:pitestRelease')
         then:
             result.wasExecuted(':itest:pitestRelease')
             result.getStandardOutput().contains('Generated 2 mutations Killed 2 (100%)')