+++ src/funcTest/resources/testProjects/simpleKotlin/src/main/kotlin/pitest/test/Counter.kt
@@ -1,0 +1,5 @@
+package pitest.test
+
+internal data class Counter(private val value: Int) {
+    fun isLessThan(otherCounter: Counter) = this.value < otherCounter.value
+}